pyproject.toml
[tool.poetry]
name = "rays"
version = "0.1.0"
description = ""
authors = ["Fuksz Domonkos <fukszdomonkos@gmail.com>"]
readme = "README.md"
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
raylib = "^5.0.0.4"
ffmpeg-python = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
